adaptive_writing_system/requirements.txt
pandas
pyyaml
scikit-learn
pgmpy
streamlit
